Many display results improvements Phase with period between means Credibility computed and displayed Better display performances for graphs (using only necessary points based on graph width in pixels) HPD computed only for chain concatenation. Display numerical results with Q1, Q2, Q3 and intervals

+++ src/mcmc/Functions.h
@@ -6,10 +6,10 @@
 
 
 struct FunctionAnalysis{
-    double max = 0.;
-    double mode = 0.;
-    double mean = 0.;
-    double variance = 0.;
+    float max = 0.;
+    float mode = 0.;
+    float mean = 0.;
+    float variance = 0.;
 };
 
 FunctionAnalysis analyseFunction(const QMap<float, float>& aFunction);
